Tags are working properly now
Before I was storing them as the user introduces them.

Now, I am removing extra white spaces and adding a comma at the
beginning and at the end. That way is much more easy to query for ,tag,
instead all the other possibilities that you could have.

It was failing before. Example:

something, django, else -> if I look for go tag it would match with the
django one.