Ceil() the number of pages
We were not doing it and it was causing that for small amount of
articles and a big number for pagination the result was close to 0 and
the index was never getting written.

Also, the number of pages was being shown on the template even when it
was just one. This behaviour was changed.